+++ main/library_tree/properties.js
@@ -326,7 +326,8 @@ let properties = [
 	['Library Source: Playback Queue force sorting', true, 'queueSorting'], // Regorxxx <- Queue source ->
 	['Library Source: Playback Queue show now playing', true, 'queueNowPlaying'], // Regorxxx <- Queue source ->
 	['Library Source: Use cache (experimental)', true, 'libSourceCache'], // Regorxxx <- Internal cache of views ->
-	['Library Source: Playlist force sorting', false, 'plsSorting'], // Regorxxx <- Internal cache of views ->
+	['Library Source: Playlist force sorting', false, 'plsSorting'], // Regorxxx <- Support playlist sorting ->
+	['Library Source: Playlist flat view', true, 'plsFlatView'], // Regorxxx <- Multiple-playlist flat view ->
 
 	['Limit Menu Expand: 10-6000', 500, 'treeExpandLimit'],
 	['Limit Tree Auto Expand: 10-1000', 350, 'autoExpandLimit'],